requirements.txt
fastapi==0.115.5
loguru==0.7.2
opencv-python==4.10.0.84
pip==24.2
python-dotenv==1.0.1
python-multipart==0.0.17
streamlit==1.40.1
uvicorn==0.32.0
